refactor: 모듈 분리 및 타입 정의 추가
