+++ client/src/components/common/Dropdown/Menu.tsx
@@ -1,0 +1,13 @@
+import { useDropdownContext } from './DropdownContext';
+
+interface MenuProps {
+  children: React.ReactNode;
+}
+
+function Menu({ children }: MenuProps) {
+  const { isOpen } = useDropdownContext();
+
+  return <>{isOpen && <div>{children}</div>}</>;
+}
+
+export default Menu;
